Support emulated assumed roles for the purpose of session tokens
The Proxy uses AccessKey/SecretKey to do AWS request auth. This
commit adds optional support for using AWS's STS IAM assume role
functionality to add session token support.

A new endpoint for STS actions is added. Currently, it only handles
`AssumeRole` action.

When an `AssumeRole` request is POSTed, the new `StsController` is
called to assume the role. How that is managed is implementation
dependent. The method must return a new set of emulated credentials
that will be returned to the calling the client. Future requests
to the S3 proxy will use those credentials.

When the S3 Proxy endpoint receives a request. It first checks
the `StsController` to see if the emulated AccessKey and Session
map to an assumed role. If so, those assumed role credentials
are used to validate the request.

Doing session tokens via an IAM emulation allows users to re-use
their existing clients, etc. with very little change. This will
integrate into most existing AWS workflows.

NOTE: this is _not_ IAM role assume for real requests to AWS. That
will be handled separately.