STS Endpoint for assuming roles
A new endpoint for STS actions is added. Currently, it only handles
`AssumeRole` action.

When an `AssumeRole` request is POSTed, the `CredentialsController` is
called to assume the role. How that is managed is implementation
dependent. The method must return a new set of emulated credentials
that will be returned to the calling the client. Future requests
to the S3 proxy will use those credentials.