Allow the remote connection to use an assumed IAM role
Adds a new `RemoteSessionRole` to `Credentials`. The assigning
and management these roles is an implementation detail of the
user of the proxy. Credentials that have remote session roles
will get automatic IAM session management via the standard AWS
assume-role feature. The assume role session is managed and
renewed internally by the proxy.

Closes #22