Add support for put/delete object
Proxies put object by piping uploaded data from the source to
the remote system. Note: the payload is sent to the remote system
unsigned. We will already have validated the request signature and
TLS/SSL makes the upload safe. If we really wanted to sign, we'd
have to do a lot more work and there doesn't appear to be any benefit.

Also adds support for delete object.

Closes #21