PySpark container, testing utils and an int test
New PySpark (Spark) container that relies on Hive Metastore
container.

PySpark is very difficult to drive from tests. We need to attach
to the container and inject `stdin`. Added a new util to help
with this and there's an int test that creates a bucket and uploads
a CSV file. Then, uses PySpark SQL to create a table backed by
the CSV file and query it.