Introduce SigningHeaders to manage which headers get signed
When validating an incoming request we can't know which version of
the AWS client is being used. We could insist on only supporting
the latest AWS clients but that would be limiting. Instead, use
the signed headers listed in the Authorization header to determine
which headers to include in the validation signature. This is secure
in that the Authorization is signed with the secret key so this list
of signed headers cannot be tampered with.

For outgoing requests, we can rely on the `AwsS3V4Signer` to do the
right thing.

Introduce SigningHeaders to make manage these distinctions easier.