Rewrite of AWS Chunked support
Rewrite of `AwsChunkedInputStream` for simplicity/clarity. In
addition when the end of a chunk is reached it reads the next
chunk's header. This ensures we don't send the final bytes of
a request before they've been validated against the signature.
i.e. the final chunk's signature cannot be validated until
the final chunk has been read. But, without this change, the
final chunk's bytes will already have been sent.