Simplify DB security
S3 database security is simplified into a decorator rather than
the previous special casing and controller. This reduces what
it was accomplishing to only what's needed: a) seeing if the S3
path maps to a table name and, if so b) basing the security response
on that table name.

We no longer have the logging that was in the controller. That can
either be resurrected in the future or pushed into clients to handle.
It likely wasn't very useful as it was anyway.